Added headers authorization for Home, workoutForm, and workoutDetails

+++ frontend/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 // hooks import
+import useAuthContext from "../hooks/useAuthContext"
 import useWorkoutsContext from "../hooks/useWorkoutsContext"
 
 // libraries import
@@ -6,15 +7,22 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 export default function WorkoutDetails({ workout }) {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
         if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
+            dispatch({ type: 'DELETE_WORKOUT', payload: json })
         }
     }
     return (
@@ -22,7 +30,7 @@ export default function WorkoutDetails({ workout }) {
             <h4>{workout.title}</h4>
             <p><strong>Load (kg): </strong>{workout.load}</p>
             <p><strong>Reps: </strong>{workout.reps}</p>
-            <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
+            <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
             <span className="material-symbols-outlined" onClick={handleClick}>Delete</span>
         </div>
     )
